+++ admin/export.js
@@ -49,7 +49,7 @@ jQuery(document).ready(function($) {
  */
 function efml_export_save_config() {
   // get all form data.
-  let formData = jQuery('.efml-export-config :input').serializeArray();
+  let formData = jQuery( '.efml-export-config :input, .efml-export-config select' ).serializeArray();
 
   // add data to process this request.
   formData.push({ 'name': 'action', 'value': 'efml_save_export_config'});
@@ -60,6 +60,7 @@ function efml_export_save_config() {
     url: efmlJsVars.ajax_url,
     type: 'POST',
     data: formData,
+    dataType: 'json',
     error: function( jqXHR, textStatus, errorThrown ) {
       efml_ajax_error_dialog( errorThrown )
     },